+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* 2026-03-24 - AB#97: Add CollectionOfTypeDescriptor and serialization test.
+* 2026-03-24 - AB#97: Make newType optional in MemberInitDescriptor constructor.
 * 2026-03-19 - AB#101: Update Codecov action and coverlet.msbuild version
 * 2036-03-16 - AB#82: Add CodeCov and testing to workflows and badges to README.md.
 * 2026-03-16 - AB#81 Removing data classes from the test project.